requirements.txt
apache-airflow
sqlparse
apache-airflow-providers-postgres 
apache-airflow-providers-docker
pandas